DEV-2675: Address review - crash-safe warning, docs, and internal helper
Makes describeValue total (String throws on a null-prototype object), moves parsePixelSize off the public helper surface into utils, mirrors the rowHeaderWidth API tags, keys the warning on the value with a per-instance scope, documents the negative number vs negative string split, corrects the 18.0.0 mechanism note, and covers mixed and non-last-entry arrays.
