+++ docs/content/guides/rows/row-header/row-header.md
@@ -141,7 +141,11 @@ To control the header size, set [`rowHeaderWidth`](@/api/options.md#rowheaderwid
 The width is a number of pixels. A string that states a pixel size works too, so a value coming from
 an attribute, a JSON config, or a framework template still applies. You can mix both forms inside the
 array. A value that states no pixel count, such as `'50%'` or `'20em'`, is ignored - the header keeps
-its default width, and Handsontable reports the value once in the browser console.
+its default width, and Handsontable reports the value once in the browser console. Inside an array
+this applies per level, so one unreadable entry does not disturb the levels around it.
+
+A negative number is kept as it is, because numbers behave exactly as they did before this option
+read strings. A negative string is rejected, so a typo cannot collapse the header.
 
 The [Row headers as an array](#row-headers-as-an-array) example uses custom labels together with `rowHeaderWidth: 80`.
