+++ handsontable/src/helpers/number.ts
@@ -330,39 +330,3 @@ export function isLossyNumericConversion(rawInput: string, parsedNumber: number)
 export function isUnsignedNumber(value: unknown): value is number {
   return Number.isInteger(value) && typeof value === 'number' && value >= 0;
 }
-
-/**
- * Matches a size written as a bare number (`'100'`, `'12.5'`) or with a `px` unit (`'100px'`),
- * ignoring surrounding whitespace and the unit's letter case.
- */
-const PIXEL_SIZE_PATTERN = /^\s*(\d+(?:\.\d+)?)\s*(?:px)?\s*$/i;
-
-/**
- * Resolves a pixel size setting written as a number or a string into a number.
- *
- * The grid's size options are documented as numbers of pixels. This helper additionally accepts the
- * two string forms that carry the same unambiguous meaning – a bare numeric string (`'100'`) and a
- * pixel string (`'100px'`) – so a value arriving from an attribute, a JSON config, or a framework
- * template still resolves. A number is returned unchanged, including a negative one, so the callers'
- * existing behavior for numbers is untouched.
- *
- * Any value that does not resolve to a pixel count returns `null`. That covers relative units
- * (`'50%'`, `'20em'`), which depend on a layout context this setting has no access to, and plain
- * text. Returning `null` lets the caller apply its own default instead of rendering a broken size.
- *
- * @param {*} value The value to resolve.
- * @returns {number|null}
- */
-export function parsePixelSize(value: unknown): number | null {
-  if (typeof value === 'number') {
-    return value;
-  }
-
-  if (typeof value !== 'string') {
-    return null;
-  }
-
-  const match = PIXEL_SIZE_PATTERN.exec(value);
-
-  return match === null ? null : parseFloat(match[1]);
-}
